// apps/demo-07/lib/mock-data/types.ts
export interface StoreInfo {
  id: string;
  name: string;
  address: string;
  phone: string;
  hours: string;
  city: string;
}